bugfix/LIVE-3251-ratings-and-notifs-modals-conflicts-fixed moved the code specific to modals in a new file
